adding sample edit icon
