+++ src/frontend/src/common/components/library/data_subview/components/MoreActionMenu/index.tsx
@@ -1,8 +1,11 @@
 import { Menu, MenuItem } from "czifui";
 import React, { MouseEventHandler, useState } from "react";
+import { StyledEditIcon, StyledTrashIcon } from "src/common/styles/iconStyle";
+import { FEATURE_FLAGS, usesFeatureFlag } from "src/common/utils/featureFlags";
+import { StyledText } from "src/views/Data/components/TreeActionMenu/components/MoreActionsMenu/style";
 import { TooltipDescriptionText, TooltipHeaderText } from "../../style";
 import { IconButton } from "../IconButton";
-import { StyledMoreActionsIcon, StyledText, StyledTrashIcon } from "./style";
+import { StyledMoreActionsIcon } from "./style";
 
 interface Props {
   disabled: boolean;
@@ -66,9 +69,15 @@ const MoreActionsMenu = ({
         onClose={handleClose}
         getContentAnchorEl={null}
       >
+        {usesFeatureFlag(FEATURE_FLAGS.editSamples) && (
+          <MenuItem onClick={() => undefined}>
+            <StyledEditIcon />
+            <StyledText>Edit Samples</StyledText>
+          </MenuItem>
+        )}
         <MenuItem onClick={handleDeleteSamples}>
           <StyledTrashIcon />
-          <StyledText>Delete Samples</StyledText>
+          <StyledText isWarning>Delete Samples</StyledText>
         </MenuItem>
       </Menu>
     </>
